requirements.txt
celery==4.0.2
Flask==0.12
gunicorn==19.7.1
nltk==3.2.2
numpy==1.12.1
python-dotenv==0.6.4
redis==2.10.5
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
